backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langgraph
langchain-openai
langchain-anthropic
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
sqlalchemy
python-dotenv
